blockchain/requirements_errors_fixes.txt
jsonpickle==1.4.1
json5==0.8.4
Flask==1.1.1
Flask-Classful==0.14.2
p2pnetwork==0.0.3
requests==2.22.0
pycrypto==2.6.1
pypubsub







-----------------------------
Errors and fixes:

Error:
AttributeError: module 'time' has no attribute 'clock'
Solution: Install pycryptodome 

Error:
AttributeError: function/symbol 'SHA256_init' not found in library
Solution: SHA1 instead of SHA256

-----------------------------
How stop localhost ports:  (Cntrl+C) Twice or the following:

https://stackoverflow.com/questions/19071512/socket-error-errno-48-address-already-in-use

Find the process using port $PORTID:
sudo lsof -I:$PORTID

Kill the process on that port:
kill $PID or PROCESSID

-----------------------------

